fix/renamed files and removed comments
