+++ nextjs/src/features/organization/connectionIssuance/components/Issuance.tsx
@@ -24,7 +24,7 @@ import FieldArrayData from './FieldArray'
 import Loader from '@/components/Loader'
 import PageContainer from '@/components/layout/page-container'
 import { RootState } from '@/lib/store'
-import { SearchableSelect } from '@/components/ShadCnSelect'
+import { SearchableSelect } from '@/components/SearchableSelect'
 import SummaryCard from '@/components/SummaryCard'
 import SummaryCardW3c from '@/components/SummaryCardW3c'
 import { apiStatusCodes } from '@/config/CommonConstant'
