+++ nextjs/src/features/organization/emailIssuance/components/EmailIssuance.tsx
@@ -41,7 +41,7 @@ import PageContainer from '@/components/layout/page-container'
 import { Plus } from 'lucide-react'
 import RoleViewButton from '@/components/RoleViewButton'
 import { RootState } from '@/lib/store'
-import { SearchableSelect } from '@/components/ShadCnSelect'
+import { SearchableSelect } from '@/components/SearchableSelect'
 import { SelectRef } from '../../bulkIssuance/components/BulkIssuance'
 import { itemPerPage } from '@/config/CommonConstant'
 import { pathRoutes } from '@/config/pathRoutes'
